feat(ci): add GitHub Actions workflow for continuous integration
- Introduced a CI workflow in `ci.yml` to automate checks for typos, Rust formatting, Clippy linting, and cargo checks on push and pull request events for the master and main branches.
- Configured jobs to run on Ubuntu with caching for improved performance.
- Ensured that the Rust toolchain is set to nightly for specific tasks and stable for others, enhancing the development workflow.

+++ pegaflow-core/src/storage.rs
@@ -320,7 +320,7 @@ impl StorageEngine {
                 let (used, total) = self.pinned_pool.usage();
                 error!(
                     "Pinned memory pool exhausted! Requested: {}, Used: {}, Total: {}, Cache empty",
-                    ByteSize(size.get() as u64),
+                    ByteSize(size.get()),
                     ByteSize(used),
                     ByteSize(total)
                 );
